% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phmm.R
\name{AIC.phmm}
\alias{AIC.phmm}
\title{Akaike Information Criterion for PHMM}
\usage{
\method{AIC}{phmm}(object, ..., k = 2)
}
\arguments{
\item{object}{an object of class \code{phmm}.}

\item{...}{optionally more fitted model objects.}

\item{k}{numeric, the penalty per parameter to be used; the default k = 2 is
the classical AIC.}
}
\value{
Returns a list of AIC values corresonding to all available
log-likelihood values from the fit. See \code{\link{phmm}} for details of
the log-likelihood values.
}
\description{
Function calculating the Akaike information criterion for PHMM fitted model
objects, according to the formula \eqn{-2*log-likelihood +
k*rho}{-2*log-likelihood + k*npar}, where \eqn{npar}{npar} represents the
number of parameters in the fitted model. The function returns a list of AIC
calculations corresponding different likelihood estimations: conditional and
marginal likelihoods calculated by Laplace approximation, reciprocal
importance sampling, and bridge sampling (only implemented for nreff < 3).
The default k = 2, is for the usual AIC.
}
\references{
Whitehead, J. (1980). Fitting Cox's Regression Model to Survival
Data using GLIM. Journal of the Royal Statistical Society. Series C, Applied
statistics, 29(3), 268-.
}
\seealso{
\code{\link{phmm}}, \code{\link[stats]{AIC}}
}
\keyword{survival}
