% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phmm.R
\name{loglik.cond}
\alias{loglik.cond}
\alias{loglik.cond.phmm}
\title{PHMM conditional log-likelihood}
\usage{
loglik.cond(x)

\method{loglik.cond}{phmm}(x)
}
\arguments{
\item{x}{an object of class \code{phmm}.}
}
\value{
The PHMM log-likelihood conditional on the estimated random effects.
}
\description{
Function for computing log-likelihood conditional on the estimated random
effects from an object of class \code{phmm} returned by \code{phmm}.
}
\seealso{
\code{\link{phmm}}, \code{\link{phmm.cond.loglik}}
}
\keyword{survival}
