% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoPoisPHMM.R
\name{pseudoPoisPHMM}
\alias{pseudoPoisPHMM}
\title{Pseudo poisson data for fitting PHMM via GLMM}
\usage{
pseudoPoisPHMM(x)
}
\arguments{
\item{x}{an object of class \code{phmm}.}
}
\value{
A \code{data.frame} with columns:
}
\description{
Function for generating a pseudo Poisson data set which can be used to fit a
PHMM using GLMM software. This follows the mixed-model extension Whitehead
(1980), who described how to fit Cox (fixed effects) models with GLM
software.
}
\references{
Whitehead, J. (1980). Fitting Cox's Regression Model to Survival
Data using GLIM. Journal of the Royal Statistical Society. Series C, Applied
statistics, 29(3).  268-.
}
\seealso{
\code{\link{phmm}}, \code{\link{traceHat}}
}
\keyword{survival}
