% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phmm.R
\name{summary.phmm}
\alias{summary.phmm}
\title{Summarizing phmm model fits}
\usage{
\method{summary}{phmm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{phmm}.}

\item{...}{not used.}
}
\description{
\code{summary} method for class \code{"phmm"}.
}

