% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceHat.R
\name{traceHat}
\alias{traceHat}
\title{Trace of the "hat" matrix from PHMM-MCEM fit}
\usage{
traceHat(x, method = "direct")
}
\arguments{
\item{x}{an object of class \code{phmm},}

\item{method}{acceptable values are \code{"direct"}, \code{"pseudoPois"}, or
\code{"HaLee"},}
}
\value{
The trace of the "hat" matrix which can be used as a measure of
complexity of the model.
}
\description{
Compute trace of the ``hat'' matrix from PHMM-MCEM fit using a direct
approximation method (Donohue, et al, submitted), an approximation via
hierarchical likelihoods (Ha et al, 2007), or an approximation via a
generalized linear mixed-effects model (GLMM) (Donohue, et al, submitted).
}
\references{
Breslow, NE, Clayton, DG. (1993). Approximate Inference in
Generalized Linear Mixed Models. Journal of the American Statistical
Association, Vol. 88, No. 421, pp.  9-25.

Donohue, M, Xu, R, Vaida, F, Haut R. Model Selection for Clustered Data:
Conditional Akaike Information under GLMM and PHMM. Submitted.

Ha, ID, Lee, Y, MacKenzie, G. (2007). Model Selection for multi-component
frailty models. Statistics in Medicine, Vol. 26, pp. 4790-4807.

Whitehead, J. (1980). Fitting Cox\'s Regression Model to Survival Data using
GLIM. Journal of the Royal Statistical Society. Series C, Applied
statistics, 29(3), 268-.
}
\seealso{
\code{\link{phmm}}, \code{\link{AIC.phmm}}
}
\keyword{survival}
